Separate controller for trip point
